back-end/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
Flask_PyMongo==2.3.0
joblib==1.4.0
pandas==2.2.2
PyJWT==2.8.0
PyJWT==2.8.0
pymongo==4.7.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
scikit_learn==1.4.2
